% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{AF2cfs.monthly}
\alias{AF2cfs.monthly}
\title{Convert a data frame of total monthly discharges in AF to average monthly
flows in cfs.}
\usage{
AF2cfs.monthly(data)
}
\arguments{
\item{data}{a data frame formatted as described above}
}
\description{
Behaves as TAF2cfs.monthly, but assumes that input are in AF.
}
\details{
This function should be rendered obsolete by \code{cfs_monthly}, but is
retained for backward compatibility with earlier scripts.
}
