% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{AF2cfs}
\alias{AF2cfs}
\title{Convert a sequence of daily discharges in AF to daily flows in cfs.}
\usage{
AF2cfs(vol, days = 1)
}
\arguments{
\item{vol}{a sequence of daily or multi-day discharges in AF}

\item{days}{(optional) number of days over which discharges are specified}
}
\description{
Convert a sequence of daily discharges in AF to daily flows in cfs.
}
