% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{boxplotAdd}
\alias{boxplotAdd}
\title{Add a second boxplot to an existing plot, formatted to be distinguishable.}
\usage{
boxplotAdd(..., border = "blue", boxwex = 0.5)
}
\arguments{
\item{...}{parameters to be passed to boxplot}

\item{border}{color for border of new boxes}

\item{boxwex}{relative width of new boxes}
}
\description{
Add a second boxplot to an existing plot, formatted to be distinguishable.
}
