% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{boxplotPaired}
\alias{boxplotPaired}
\title{Draw a paired boxplot}
\usage{
boxplotPaired(formula, df1, df2, border = c("black", "blue"),
  width = 0.25, sepFactor = 1.5, ...)
}
\arguments{
\item{formula}{formula of form y~x to be applied to both data frames}

\item{df1}{data frame with values to be plotted at left}

\item{df2}{data frame with values to be plotted at right}

\item{border}{vector of colors for borders of boxes}

\item{width}{relative width of plots}

\item{sepFactor}{relative separation between boxes in a pair [describe better]}

\item{...}{additional paramaters to be passed to boxplot}
}
\description{
Draw a paired boxplot
}
