% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{cfs2TAF.monthly}
\alias{cfs2TAF.monthly}
\title{Convert a data frame of monthly flows in cfs to total monthly volume in TAF.
The data frame should consist of monthly average flows and contain a column
named "date" of class POSIXt or Date. All columns of class "numeric" are
taken to be monthly average flows in cfs. All other columns beside "date" are
discarded.}
\usage{
cfs2TAF.monthly(data)
}
\arguments{
\item{data}{a data frame formatted as described above}
}
\description{
This function should be rendered obsolete by \code{taf_monthly}, but is
retained for backward compatibility with earlier scripts.
}
