% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{cfs2TAF}
\alias{cfs2TAF}
\title{Convert a sequence of daily flows in cfs to discharges in TAF.}
\usage{
cfs2TAF(flow, days = 1)
}
\arguments{
\item{flow}{a sequence of flow rates in cfs}

\item{days}{(optional) number of days over which to accumulate flows}
}
\description{
Convert a sequence of daily flows in cfs to discharges in TAF.
}
