% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{cleanWEAPHeader}
\alias{cleanWEAPHeader}
\title{Clean a header from a ReadFromFile csv format file exported from WEAP.}
\usage{
cleanWEAPHeader(header)
}
\arguments{
\item{header}{the raw or minimally processed header}
}
\description{
Clean a header from a ReadFromFile csv format file exported from WEAP.
}
