% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dayflow}
\alias{dayflow}
\title{Full Dayflow data set for water years 1956-2017.}
\format{
A data frame with 22646 rows and 29 columns:
\describe{
 \item{year}{calendar year}
 \item{mo}{month (integer)}
 \item{date}{date (Date)}
 \item{sac}{Sacramento River flow at Freeport}
 \item{yolo}{Yolo Bypass at Woodland}
 \item{csmr}{Cosumnes River}
 \item{moke}{Mokelumne River}
 \item{misc}{Miscellaneous streams}
 \item{sjr}{San Joaquin River at Vernalis}
 \item{east}{Eastern Delta inflow}
 \item{tot}{Total Delta inflow}
 \item{ccc}{Contra Costa Water District diversions}
 \item{swp}{State Water Project exports}
 \item{cvp}{Central Valley Project exports}
 \item{nbaq}{North Bay Aqueduct exports}
 \item{out}{Net Delta outflow}
 \item{exports}{Total Delta exports}
 \item{gcd}{Delta gross channel depletions}
 \item{prec}{Delta precipitation}
 \item{misdv}{Miscellaneous Delta diversions}
 \item{cd}{Delta net channel depletions}
 \item{xgeo}{Delta Cross Channel plus Georgiana Slough}
 \item{west}{San Joaquin River at Jersey Point}
 \item{rio}{Sacramento River at Rio Vista}
 \item{expin}{Export/Inflow ratio}
 \item{diver}{Percent water diverted from the Delta}
 \item{effec}{Effective inflow to the western/central Delta}
 \item{effdiv}{Effective percent diverted from the western/central Delta}
 \item{x2}{Estimated distance from Golden Gate to 2 ppt salinity, km}
}
}
\source{
\url{https://water.ca.gov/Programs/Environmental-Services/Compliance-Monitoring-And-Assessment/Dayflow-Data}
}
\usage{
dayflow
}
\description{
Most values are average daily net flows in cubic feet per second. Full
documentation available at
\url{https://water.ca.gov/-/media/DWR-Website/Web-Pages/Programs/Environmental-Services/Compliance-Monitoring--Assessment/Dayflow/Files/Publications/Current-Dayflow-Documentation.pdf}.
}
\details{
Last updated July 16, 2018.
}
\keyword{datasets}
