% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{eightRiverMonthly}
\alias{eightRiverMonthly}
\title{Return a data frame containing the monthly full natural flow volumes for the
Eight River Index stations in acre-feet.}
\usage{
eightRiverMonthly(startDate = as.Date("1905-10-01"),
  endDate = Sys.Date(), raw = FALSE)
}
\arguments{
\item{startDate}{start date for time series}

\item{endDate}{end date for time series}

\item{raw}{if TRUE, return a data frame with monthly values for each station;
if FALSE, return a data frame containing only the summed index}
}
\description{
Return a data frame containing the monthly full natural flow volumes for the
Eight River Index stations in acre-feet.
}
