% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eomDate.character}
\alias{eomDate.character}
\title{S3 method implementing \code{eomDate} for character strings in ymd format.}
\usage{
\method{eomDate}{character}(date)
}
\arguments{
\item{date}{date string in ymd format}
}
\description{
S3 method implementing \code{eomDate} for character strings in ymd format.
}
