% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eomDate.Date}
\alias{eomDate.Date}
\title{S3 method implementing \code{eomDate} for dates specified with class Date.}
\usage{
\method{eomDate}{Date}(date)
}
\arguments{
\item{date}{date of class Date}
}
\description{
S3 method implementing \code{eomDate} for dates specified with class Date.
}
