% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eomDate.POSIXt}
\alias{eomDate.POSIXt}
\title{S3 method implementing \code{eomDate} for dates specified with class POSIXt.}
\usage{
\method{eomDate}{POSIXt}(date)
}
\arguments{
\item{date}{date or datetime of class POSIXt}
}
\description{
S3 method implementing \code{eomDate} for dates specified with class POSIXt.
}
