% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eomDate_wym}
\alias{eomDate_wym}
\title{Return the date of the last day of the month in a given water year, month
pair.}
\usage{
eomDate_wym(wy, month, wyStart = 10)
}
\arguments{
\item{wy}{water year as integer}

\item{month}{month as integer or string from \code{month.abb}}

\item{wyStart}{(optional) first month in the water year as integer, if
different from California convention of October}
}
\description{
Return the date of the last day of the month in a given water year, month
pair.
}
