% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eomDate_ym}
\alias{eomDate_ym}
\title{Return the date of the last day in the month for a given year, month pair.}
\usage{
eomDate_ym(year, month)
}
\arguments{
\item{year}{year as integer}

\item{month}{month as integer or string from \code{month.abb}}
}
\description{
Return the date of the last day in the month for a given year, month pair.
}
