% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eoqDate.character}
\alias{eoqDate.character}
\title{S3 method implementing \code{eoqDate} for character strings in ymd format.}
\usage{
\method{eoqDate}{character}(date)
}
\arguments{
\item{date}{date string in ymd format}
}
\description{
S3 method implementing \code{eoqDate} for character strings in ymd format.
}
