% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eoqDate.Date}
\alias{eoqDate.Date}
\title{S3 method implementing \code{eoqDate} for dates specified with class Date.}
\usage{
\method{eoqDate}{Date}(date)
}
\arguments{
\item{date}{date of class Date}
}
\description{
S3 method implementing \code{eoqDate} for dates specified with class Date.
}
