% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eoqDate.POSIXt}
\alias{eoqDate.POSIXt}
\title{S3 method implementing \code{eoqDate} for dates specified with class POSIXt.}
\usage{
\method{eoqDate}{POSIXt}(date)
}
\arguments{
\item{date}{date or datetime of class POSIXt}
}
\description{
S3 method implementing \code{eoqDate} for dates specified with class POSIXt.
}
