% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{eoqDate}
\alias{eoqDate}
\title{Return the date of the last day in the quarter for a given date.}
\usage{
eoqDate(date)
}
\arguments{
\item{date}{date in POSIXt, Date, or ymd string}
}
\description{
This is a generic function that dispatches the appropriate S3 method.
Operates on scalar or vector arguments.
}
