% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{excf}
\alias{excf}
\title{Return a closure that calculates the empirical exceedance function.}
\usage{
excf(x)
}
\arguments{
\item{x}{numeric vector with one or more non-missing values}
}
\description{
Modified from stats::ecdf, which calculates the empirical cumulative
distribution function.
}
