% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_schedule.R
\name{flow_schedule}
\alias{flow_schedule}
\title{An S3 class to represent flow schedules on a fixed time step (defaults
to monthly). This class is under development and should be used with caution.}
\usage{
flow_schedule(flow, time, return_freq = 0.5, step = "month",
  wy_start = 10)
}
\arguments{
\item{flow}{a sequence of flows}

\item{time}{a sequence of time steps}

\item{return_freq}{desired return frequency for the flow schedule, if any}

\item{step}{units of time step}

\item{wy_start}{first month of water year (integer)}
}
\description{
An S3 class to represent flow schedules on a fixed time step (defaults
to monthly). This class is under development and should be used with caution.
}
