% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{getCDECMulti}
\alias{getCDECMulti}
\title{Retrieve data from CDEC for multiple sensors at multiple stations.}
\usage{
getCDECMulti(stations, sensors, duration = "D",
  startDate = "1900-10-01", endDate = Sys.Date())
}
\arguments{
\item{stations}{n-element vector of station IDs}

\item{sensors}{n-element vector of sensor IDs}

\item{duration}{common duration code}

\item{startDate}{start date for time series}

\item{endDate}{end date for time series}
}
\description{
Retrieve data from CDEC for multiple sensors at multiple stations.
}
