% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{getDayflow}
\alias{getDayflow}
\title{Downloads and compile the Dayflow data set from
http://www.water.ca.gov/dayflow/output/Output.cfm.}
\usage{
getDayflow(files)
}
\arguments{
\item{files}{list of files in form list(merged=list(), single=list())}
}
\description{
Downloads and compile the Dayflow data set from
http://www.water.ca.gov/dayflow/output/Output.cfm.
}
