% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_schedule.R
\name{length.flow_schedule}
\alias{length.flow_schedule}
\title{Override the default behavior of \code{length} for \code{flow_schedule}
objects, which are implemented as lists, to prevent them returning the
number of elements in the object itself.}
\usage{
\method{length}{flow_schedule}(x)
}
\arguments{
\item{x}{\code{flow_schedule} object}
}
\description{
TODO: Evaluate whether this is the correct solution. To this point, we have
used lists to contain multiple \code{flow_schedule} objects, but there is no
reason that we could not use vectors. This implementation simply returns
\code{1}, which would not be the correct behavior for a vector.
}
