% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_schedule.R
\name{mean.flow_schedule}
\alias{mean.flow_schedule}
\title{Implement an S3 method for \code{mean} for \code{flow_schedule}}
\usage{
\method{mean}{flow_schedule}(x)
}
\arguments{
\item{x}{object of class \code{flow_schedule}}
}
\description{
Implement an S3 method for \code{mean} for \code{flow_schedule}
}
