% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{min_max_threshold}
\alias{min_max_threshold}
\title{Return a sequence of values constrained within an inclusive minimum and
maximum value.}
\usage{
min_max_threshold(x, a, b)
}
\arguments{
\item{x}{sequence of orderable values}

\item{a}{minimum value}

\item{b}{maximum value}
}
\description{
Return a sequence of values constrained within an inclusive minimum and
maximum value.
}
