% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{names_rm_args}
\alias{names_rm_args}
\title{Remove parenthetical arguments from the names of an object containing named
elements (e.g., a data frame).}
\usage{
names_rm_args(x)
}
\arguments{
\item{x}{object with names that need to be cleaned up}
}
\description{
This arises most commonly when a data frame is created with a variable that
is the result of a function call.
}
