% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{path_collapse}
\alias{path_collapse}
\title{Collapse a sequence of arguments into a file path.}
\usage{
path_collapse(...)
}
\arguments{
\item{...}{two or more strings to be collapsed into a file path}
}
\description{
This is purely a convenience utility, and is equivalent to
\code{paste(..., sep="/")}.
}
