% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotExceedance}
\alias{plotExceedance}
\title{Plot an exceedance curve.}
\usage{
plotExceedance(x, add = F, pch = 19, col = "black",
  xlab = "Exceedance Frequency", ylab = "x", ylim = NULL,
  xlim = c(0, 1), bg = NA, type = "p", jitter = FALSE,
  jitterTol = 1e-09, plot.pos = "center", ...)
}
\arguments{
\item{x}{the data series}

\item{add}{if TRUE, add plot to existing axes}

\item{pch}{plotting character}

\item{col}{color}

\item{xlab}{xlabel text}

\item{ylab}{ylabel text}

\item{ylim}{y-axis limits}

\item{xlim}{x-axis limits}

\item{bg}{background color}

\item{type}{one of \code{p} (points), \code{l} (line), \code{b} (both), or
\code{n} (none)}

\item{jitter}{if TRUE, jitter points randomly to prevent overplotting}

\item{jitterTol}{relative magnitude for jittering points}

\item{plot.pos}{one of \code{left} (maximum plotted at 0), \code{right}
(minimum plotted at 1), or \code{center} (positions centered between extremes
described above); the distance between plotting positions is always
\code{1/n}, provided that the values in \code{x} are distinct (or points are
jittered)}

\item{...}{additional parameters passed to \code{plot}}
}
\description{
Plot an exceedance curve with plotting positions determined by the empirical
cumulative distribution function.
}
\details{
Note: should consider modifying to use Weibull plotting positions or
providing more choices.
}
