% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotExceedanceComparison}
\alias{plotExceedanceComparison}
\title{Plot exceedance curves for each series of values in a list.}
\usage{
plotExceedanceComparison(l.x, v.col = NULL, ylim = NULL, main = NULL,
  ylab = NULL, xlab = "Exceedance Frequency", legend = NULL,
  legend.cex = 1, type = "p", ...)
}
\arguments{
\item{l.x}{list of data series}

\item{v.col}{vector of colors for plotting series; in NULL, a sensible
default of black and blue, or black and a rainbow gradient will be used}

\item{ylim}{y-axis limits}

\item{main}{main title for figure}

\item{ylab}{y-axis label text}

\item{xlab}{x-axis label text}

\item{legend}{vector of legend entries}

\item{legend.cex}{text scaling factor for legend}

\item{type}{one of \code{p} (points), \code{l} (line), \code{b} (both), or
\code{n} (none)}

\item{...}{additional parameters passed to \code{plotExceedance}}
}
\description{
Plot exceedance curves for each series of values in a list.
}
