% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTSComparison}
\alias{plotTSComparison}
\title{Plot multiple time series of the same pair of variables from a list of data
frames with conforming entries.}
\usage{
plotTSComparison(
  formula,
  l.df,
  v.col = NULL,
  main = NULL,
  ylab = NULL,
  xlab = "Date",
  ...
)
}
\arguments{
\item{formula}{plotting formula of form y~x}

\item{l.df}{list of data frames containing data to be plotted}

\item{v.col}{vector of colors for plotting series; in NULL, a sensible
default of black and blue, or black and a rainbow gradient will be used}

\item{main}{main title of plot}

\item{ylab}{y-axis label}

\item{xlab}{x-axis label}

\item{...}{additional parameters passed to plot}
}
\description{
Note: there is no obvious reason for this function to be used only for time
series since it is really just an xy scatter plot. Consider renaming to be
more descriptive.
}
