% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read.xlsx_sheet}
\alias{read.xlsx_sheet}
\title{Adapter function for reading multiple XLSX sheets using lapply.}
\usage{
read.xlsx_sheet(sheet, ...)
}
\arguments{
\item{sheet}{integer index or name of sheet to read}

\item{...}{additional arguments passed to read.xlsx}
}
\description{
TODO: clarify documentation.
}
