% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{readExportedC2VSim}
\alias{readExportedC2VSim}
\title{Read a table of C2VSim results converted to reach-per-column format from the
default.}
\usage{
readExportedC2VSim(file, sheet, short.names = TRUE)
}
\arguments{
\item{file}{name of XLSX file containing C2VSim results}

\item{sheet}{integer index or name of sheet to read}

\item{short.names}{if TRUE, shorten names to form "REACH.xy"; if FALSE,
retain longer names}
}
\description{
Read a table of C2VSim results converted to reach-per-column format from the
default.
}
