% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{readExportedDSS}
\alias{readExportedDSS}
\title{Import DSS data from an Excel (XLSX) file.}
\usage{
readExportedDSS(file, sheet = 1, dataRow = 8, nameRow = 2, debug = FALSE)
}
\arguments{
\item{file}{XLSX file containing one or more worksheets exported from a DSS
file}

\item{sheet}{integer index or name of sheet to read}

\item{dataRow}{first row that contains data}

\item{nameRow}{row to use for column names}

\item{debug}{if TRUE, print debugging messages}
}
\description{
Import DSS data from an Excel (XLSX) file.
}
