% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{readExportedWEAP}
\alias{readExportedWEAP}
\title{Import flow data exported from WEAP in the ReadFromFile csv format.}
\usage{
readExportedWEAP(file, divisor = 1, discardCA = TRUE, endYear = NULL,
  include.sum = FALSE, data.dir = NULL, date.only = FALSE,
  skip = 5, narrow = FALSE)
}
\arguments{
\item{file}{the full path to the file to read, or the relative path if
data.dir is set to a base path}

\item{divisor}{(optional) divisor for all flow or volume variables (e.g.,
if you want to convert data from AF to TAF)}

\item{discardCA}{whether or not to discard the current accounts year}

\item{endYear}{an optional year at which to truncate the data}

\item{include.sum}{whether or not to include a sum of all flow or volume
columns if absent}

\item{data.dir}{optional base path on the filesystem}

\item{date.only}{whether to exclude "timestep" and "year" columns from the
returned data frame}

\item{skip}{(optional) number of lines to skip before the header row}

\item{narrow}{(optional) if TRUE, convert data frame to narrow format}
}
\description{
Import flow data exported from WEAP in the ReadFromFile csv format.
}
