% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{scale_ts}
\alias{scale_ts}
\title{Scale a data frame containing a time series in which all observations are
numeric, and any time index columns are represented as a type other than
numeric (e.g., integer, date, or POSIXt).}
\usage{
scale_ts(a, x)
}
\arguments{
\item{a}{scaling factor}

\item{x}{data frame containing values to scale}
}
\description{
Scale a data frame containing a time series in which all observations are
numeric, and any time index columns are represented as a type other than
numeric (e.g., integer, date, or POSIXt).
}
