% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{summary_quantile}
\alias{summary_quantile}
\title{Calculate a cumulative distribution on specified quantiles and mean of a
numeric vector.}
\usage{
summary_quantile(x, q, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{q}{vector of quantiles to calculate}

\item{na.rm}{if TRUE, remove NA values}
}
\description{
This function is more limited than the general \code{summary} function, and
is primarily useful for generating summary tables for multiple variables
when used with \code{lapply}, as, \code{lapply(df[,cols], summary.quantile)},
where \code{cols} describes columns of a data frame containing variables to
be compared.
}
