% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svufm_ufc}
\alias{svufm_ufc}
\title{Sacramento Valley Unimpaired Flow Model data at designated locations.}
\format{A data frame with 50688 rows and 9 columns:
\describe{
 \item{date}{date (Date)}
 \item{variable}{location of the unimpaired flow estimate (DSS part B)}
 \item{value}{unimpaired flow}
 \item{unit}{unit of measurement}
 \item{a}{DSS part A}
 \item{c}{DSS part C}
 \item{e}{DSS part E}
 \item{f}{DSS part F}
 \item{type}{DSS TYPE, e.g., "PER-AVER"}
}}
\source{
\url{https://d3.water.ca.gov/owncloud/index.php/s/VHQUbo2vlhxfScG}
}
\usage{
svufm_ufc
}
\description{
[TODO: complete documentation] Last updated January 13, 2017.
}
\keyword{datasets}
