% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{TAF2cfs.monthly}
\alias{TAF2cfs.monthly}
\title{Convert a data frame of total monthly discharges in TAF to average monthly
flows in cfs.}
\usage{
TAF2cfs.monthly(data)
}
\arguments{
\item{data}{a data frame formatted as described above}
}
\description{
Behaves as cfs2TAF.monthly, but as its inverse.
}
\details{
This function should be rendered obsolete by \code{cfs_monthly}, but is
retained for backward compatibility with earlier scripts.
}
