% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uf_mapping}
\alias{uf_mapping}
\title{Unimpaired flow location mapping.}
\format{
A data frame with 23 rows and 4 columns:
\describe{
 \item{name}{descriptive name for each location}
 \item{c2vsim}{Reach designation for C2VSim coarse-grid model}
 \item{svufm}{SVUFM location name}
 \item{sacwam}{SacWAM location name}
 }
}
\usage{
uf_mapping
}
\description{
Map a subset of common unimpaired flow locations among C2VSim, SVUFM, and
SacWAM. Last updated January 13, 2017.
}
\keyword{datasets}
