% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_schedule.R
\name{unlist.flow_schedule}
\alias{unlist.flow_schedule}
\title{Override the default behavior of \code{unlist} for \code{flow_schedule}
objects, which are themselves lists, to prevent them from being
unintentionally broken down into their components.}
\usage{
\method{unlist}{flow_schedule}(x, ...)
}
\arguments{
\item{x}{object of class \code{flow_schedule}}

\item{...}{[TODO: Document why this is here; maybe just the signature of
\code{unlist}?]}
}
\description{
TODO: Document the circumstances where this arises and evaluate whether this
is really the correct solution.
}
