% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{waterYear}
\alias{waterYear}
\title{Return the water year for a date or vector of dates.}
\usage{
waterYear(date, wyStart = 10, offset = 1)
}
\arguments{
\item{date}{in Date or POSIXt format (scalar or vector).}

\item{wyStart}{(optional) first month in the water year as integer, if
different from California convention of October}

\item{offset}{(optional) signed integer offset between the calendar year
and water year in the first month of the water year; set equal to 0 if
the water year begins during the calendar year}
}
\description{
Dates must be in one of the date formats supported by lubridate.
}
