% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{wyMonth.character}
\alias{wyMonth.character}
\title{S3 method implementing \code{wyMonth} for character strings (ymd format or
long or short name of month).}
\usage{
\method{wyMonth}{character}(date, wyStart = 10)
}
\arguments{
\item{date}{date string in ymd format or month name (element of month.abb or
month.name from R base)}

\item{wyStart}{(optional) first month in the water year as integer, if
different from California convention of October}
}
\description{
Vector arguments must be consistent in format. Do not mix ymd strings with
month names from month.abb or month.name. Do not mix names from month.abb
with month.name. Month names are case insensitive, however.
}
