% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{wyMonth.Date}
\alias{wyMonth.Date}
\title{S3 method implementing \code{wyMonth} for dates in Date format.}
\usage{
\method{wyMonth}{Date}(date, wyStart = 10)
}
\arguments{
\item{date}{date in Date format}

\item{wyStart}{(optional) first month in the water year as integer, if
different from California convention of October}
}
\description{
S3 method implementing \code{wyMonth} for dates in Date format.
}
