% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{wyMonth.integer}
\alias{wyMonth.integer}
\title{S3 method implementing \code{wyMonth} for months specified as integers.}
\usage{
\method{wyMonth}{integer}(date, wyStart = 10)
}
\arguments{
\item{date}{month as integer}

\item{wyStart}{(optional) first month in the water year as integer, if
different from California convention of October}
}
\description{
S3 method implementing \code{wyMonth} for months specified as integers.
}
