% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{wyMonth.POSIXt}
\alias{wyMonth.POSIXt}
\title{S3 method implementing \code{wyMonth} for dates in POSIXt format.}
\usage{
\method{wyMonth}{POSIXt}(date, wyStart = 10)
}
\arguments{
\item{date}{date in Date format}

\item{wyStart}{(optional) first month in the water year as integer, if
different from California convention of October}
}
\description{
S3 method implementing \code{wyMonth} for dates in POSIXt format.
}
