% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{wyMonth}
\alias{wyMonth}
\title{Return the month from one or more dates as a factor in water year order.}
\usage{
wyMonth(date, wyStart = 10)
}
\arguments{
\item{date}{in Date, POSIXt, character (ymd or name of month), or integer
(month) format (scalar or vector).}

\item{wyStart}{(optional) first month in the water year as integer, if
different from California convention of October}
}
\description{
This is a generic function that dispatches the appropriate S3 method.
Operates on scalar or vector arguments.
}
\details{
Default order is "Oct", "Nov", ..., "Sep" per the convention in use in
California.
}
