% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{eightRiverDaily}
\alias{eightRiverDaily}
\title{Return a data frame containing the daily sum of full natural flows for Eight
River Index stations in cubic feet per second.}
\usage{
eightRiverDaily(startDate = as.Date("1999-10-01"),
  endDate = Sys.Date(), raw = FALSE)
}
\arguments{
\item{startDate}{start date for time series}

\item{endDate}{end date for time series}

\item{raw}{if TRUE, return a data frame with daily values for each station;
if FALSE, return a data frame containing only the summed index}
}
\description{
Return a data frame containing the daily sum of full natural flows for Eight
River Index stations in cubic feet per second.
}
