% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{readExportedDSSFull}
\alias{readExportedDSSFull}
\title{Import DSS data from an Excel (XLSX) file, retaining full metadata.}
\usage{
readExportedDSSFull(file, sheet = 1, namePart = "B")
}
\arguments{
\item{file}{XLSX file containing one or more worksheets exported from a DSS
file}

\item{sheet}{integer index or name of sheet to read}

\item{namePart}{which of the DSS parts (A-F) to use as variable names}
}
\description{
Import DSS data from an Excel (XLSX) file, retaining full metadata.
}
